+++ classes/dispatch/connection.php
@@ -1,0 +1,11 @@
+<?php defined('SYSPATH') or die('No direct script access.');
+/**
+ * Dispatch Connection
+ * 
+ * @package		Dispatch
+ * @category	Base
+ * @author		Micheal Morgan <micheal@morgan.ly>
+ * @copyright	(c) 2011 Micheal Morgan
+ * @license		MIT
+ */
+class Dispatch_Connection extends Kohana_Dispatch_Connection {}